Fix folder structure
-Arrange Folder Structure
-Reusable useFamily Data
