+++ src/components/Family/EditParent.jsx
@@ -11,7 +11,6 @@ import {
   FormMessage,
 } from "@/components/ui/form";
 import { Input } from "@/components/ui/input";
-import { z } from "zod";
 import {
   Dialog,
   DialogClose,
@@ -23,19 +22,8 @@ import {
   DialogTrigger,
 } from "@/components/ui/dialog";
 import { Button } from "@/components/ui/button";
-import { useEditParent } from "@/hooks/useUpdateFamily";
-
-const editParentSchema = z.object({
-  firstName: z.string().min(2, {
-    message: "First name must be at least 2 characters.",
-  }),
-  lastName: z.string().min(2, {
-    message: "Last name must be at least 2 characters.",
-  }),
-  contactNumber: z.string().regex(/^[0-9]{11}$/, {
-    message: "Contact number must be exactly 11 digits.",
-  }),
-});
+import { useEditParent } from "@/hooks/useFamily";
+import { editParentSchema } from "@/zodSchema/Family/EditParentSchema";
 
 const EditParent = ({
   parentId,
