+++ src/components/Home/Profile-Registration/FamilyRegistration.jsx
@@ -14,8 +14,8 @@ import { addFamilySchema } from "@/zodSchema/AddFamilySchema";
 import { Label } from "@/components/ui/label";
 import { useToast } from "@/hooks/use-toast";
 import { DialogClose } from "../../ui/dialog";
-import { useAddFamily } from "@/hooks/useAddFamily";
 import { useUser } from "@/context/useUser"; // Custom hook for accessing UserContext
+import { useAddFamily } from "@/hooks/useFamily";
 
 const FamilyRegistration = ({ skipBtn }) => {
   const { toast } = useToast();
