+++ src/pages/Dashboard.jsx
@@ -1,16 +1,4 @@
-import useUserMinistries from "@/hooks/useUserMinistries";
-
 const Dashboard = () => {
-  const { error, isLoading } = useUserMinistries(); // Use the hook
-
-  if (isLoading) {
-    return <div>Loading...</div>;
-  }
-
-  if (error) {
-    return <div>Error: {error.message}</div>;
-  }
-
   return <div className="text-heading font-bold text-accent">Dashboard</div>;
 };
