+++ src/zodSchema/Family/EditChildSchema.js
@@ -1,0 +1,10 @@
+import { z } from "zod";
+
+export const editChildSchema = z.object({
+  firstName: z.string().min(2, {
+    message: "First name must be at least 2 characters.",
+  }),
+  lastName: z.string().min(2, {
+    message: "Last name must be at least 2 characters.",
+  }),
+});
